Add synchronization of transaction values
Modifying either the Dash or USD value should update the other field.
The transaction total should also update based on changes to the "Amount
per Wallet" and "Per Transaction Fee". Currently only works with amount
field; need to discuss whether the transaction field should be updated
based on the estimate available, or editable. Also need to determine if
the Total should be editable.